refactor: 용근님 리뷰사항 적용 수정
- PathFinder 클래스 내부에 외부 라이브러리 종속시키지 않고, 인터페이스로 빼고 별도 구현체로 구현
