+++ src/main/java/nextstep/subway/path/application/PathService.java
@@ -3,6 +3,7 @@ package nextstep.subway.path.application;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.LineRepository;
 import nextstep.subway.path.domain.PathFinder;
+import nextstep.subway.path.domain.PathFinderUsingWeightedMultigraph;
 import nextstep.subway.path.domain.ShortestPath;
 import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.domain.Station;
@@ -33,7 +34,7 @@ public class PathService {
                 .orElseThrow(NoSuchElementException::new);
         List<Line> lines = lineRepository.findAll();
 
-        PathFinder pathFinder = new PathFinder(lines);
+        PathFinder pathFinder = new PathFinderUsingWeightedMultigraph(lines);
         ShortestPath shortestPath = pathFinder.executeDijkstra(sourceStation, targetStation);
 
         return PathResponse.of(shortestPath);
